Suppress harmless g++ warnings about converting float/double to int.
The conversion is done on purpose on these places so just suppress the warning
by using explicit casts.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@62679 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/richtext/richtexthtml.cpp
@@ -322,7 +322,7 @@ void wxRichTextHTMLHandler::BeginParagraphFormatting(const wxTextAttr& WXUNUSED(
             str << wxT(">");
 
             // TODO: convert to pixels
-            int indentPixels = indentLeftMM*10/4;
+            int indentPixels = static_cast<int>(indentLeftMM*10/4);
 
             if ((GetFlags() & wxRICHTEXT_HANDLER_USE_CSS) == 0)
             {
